Fix [ 685603 ] warning: cast to pointer from integer of different size and [ 654048 ] patch: incorrect cast from/to pointers. Get build on Alpha / 64-bit compilers working.  Checked that changes so far are backward compatible on 32-bit.

+++ tidy/include/fileio.h
@@ -30,7 +30,7 @@ void freeFileSource( TidyInputSource* source, Bool closeIt );
 void initFileSink( TidyOutputSink* sink, FILE* fp );
 
 /* Needed for internal declarations */
-void filesink_putByte( uint sinkData, byte bv );
+void filesink_putByte( ulong sinkData, byte bv );
 
 #ifdef __cplusplus
 }
